fix(wl): preserve collapsed surfaces and clear strict Clippy warnings
  Keep compositor-managed collapsed surfaces from being auto-resurrected
  during dock maintenance, carry updates, and pan/focus restoration, while
  also cleaning up the workspace so it passes strict warning-free checks.

  Behavior changes:
  - add a shared `preserve_collapsed_surface` helper in wl state to treat
    both manually-collapsed nodes and already-collapsed surface nodes as
    sticky
  - use that helper across carry, focus, and maintenance paths so docked
    pair geometry can still be enforced without re-heating or reactivating
    collapsed members
  - prevent pan-return focus restoration and offscreen dock decay paths from
    reviving surfaces that should remain collapsed
  - keep promote/minimize paired-surface handling intact while simplifying
    the partner gating logic

  Cleanup and warning fixes:
  - add missing `Default` impls for `FocusRingDecayPolicy`, `WeightModel`,
    and `ClientState`
  - collapse nested `if` / `if let` chains across config parsing, input,
    state, interaction, focus, maintenance, workspace, and surface lifecycle
    code
  - remove unused vector helpers from surface lifecycle code
  - replace needless borrows, unnecessary casts, `into_iter()` on borrowed
    collections, and manual clamp patterns with Clippy-preferred forms
  - simplify cursor theme loading/cache checks and a few render clipping
    branches
  - add narrow `#[allow(...)]` annotations only where existing API shape is
    intentional for now, such as wide render/input entry points, complex
    render tuple returns, spawn bookkeeping placeholders, and the
    layer-shell test-module ordering

  Verification:
  - `cargo check`
  - `cargo clippy --workspace --all-targets -- -D warnings`

+++ crates/halley-wl/src/run/winit_backend.rs
@@ -225,7 +225,7 @@ pub(super) fn run_winit_backend() -> Result<(), Box<dyn Error>> {
 
             {
                 let ws = backend.borrow().window_size();
-                publish_winit_output_snapshot(ws.w as i32, ws.h as i32, true, 0, 0);
+                publish_winit_output_snapshot(ws.w, ws.h, true, 0, 0);
             }
 
             let mut ev: EventLoop<HalleyWlState> = EventLoop::try_new()?;
@@ -466,7 +466,7 @@ pub(super) fn run_winit_backend() -> Result<(), Box<dyn Error>> {
 
                 {
                     let ws = backend_for_output_timer.borrow().window_size();
-                    publish_winit_output_snapshot(ws.w as i32, ws.h as i32, true, 0, 0);
+                    publish_winit_output_snapshot(ws.w, ws.h, true, 0, 0);
                 }
 
                 {
